separated convertions into utils.go
